Update main.py
This update adds encryption to the handler. Encryption is custom and thought out. Uses python built-in dependencies. Based off of the random class for complex output, and time.time for different outputs each run. 

Some additional updates will be coming towards these classes, but final revisions are coming soon to it.